Refactor SQL checker (#9815)
